+++ imaginenc/__init__.py
@@ -1,0 +1,10 @@
+"""Convert any file into an image"""
+
+from .imaginenc import (
+    encode_file_name,
+    encode_bytes_to_image,
+    encode_bytes_to_colors,
+    colors_to_image,
+    decode_image_name,
+    decode_image_to_bytes
+)
