+++ imaginenc/imaginenc.py
